requirements.txt
flask 
requests 
faiss-cpu 
sentence-transformers 
llama-cpp-python